refactor: move template to a separate file
